# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Main Z11 api",
        "description": "This is our main Z11 api. It contains all method to handle Z11 project...",
        "termsOfService": "",
        "contact": {
            "email": "kienlv58@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/chapter/get/{id}": {
            "get": {
                "tags": [
                    "5.Chapter"
                ],
                "summary": "get chapter",
                "description": "get chapter with chapter_id",
                "operationId": "getchapter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "chapter_id",
                        "in": "formData",
                        "description": "chapter_",
                        "required": true,
                        "schema": {
                            "required": [
                                "grant_type"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get succes"
                    },
                    "400": {
                        "description": "Invalid param"
                    }
                }
            }
        },
        "/chapter/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "5.Chapter"
                ],
                "summary": "get all Chapter",
                "description": "return chapter with take and skip",
                "operationId": "Chapter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
        "/chapter/add": {
            "post": {
                "tags": [
                    "5.Chapter"
                ],
                "summary": "add new chapter",
                "description": "add new chapter",
                "operationId": "packageadd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "package_id",
                        "in": "formData",
                        "description": "package_id ",
                        "required": true,
                        "schema": {
                            "required": [
                                "package_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "translate",
                        "in": "formData",
                        "description": "translate json",
                        "required": true,
                        "schema": {
                            "required": [
                                "category_code"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/chapter/edit": {
            "post": {
                "tags": [
                    "5.Chapter"
                ],
                "summary": "edit a Chapter",
                "description": "edit Chapter",
                "operationId": "Chapteredit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "chapter_id",
                        "in": "formData",
                        "description": "chapter_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "chapter_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "translate",
                        "in": "formData",
                        "description": "translate json",
                        "required": true,
                        "schema": {
                            "required": [
                                "category_code"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/chapter/delete": {
            "post": {
                "tags": [
                    "5.Chapter"
                ],
                "summary": "delete chapter ",
                "description": "delete with chapter_id",
                "operationId": "chapterdelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "formData",
                        "description": "uid delete",
                        "required": true,
                        "schema": {
                            "required": [
                                "uid"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "chapter_id",
                        "in": "formData",
                        "description": "chapter_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "chapter_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
        "/group_question/delete": {
            "post": {
                "tags": [
                    "6.GroupQuestion"
                ],
                "summary": "delete group_question ",
                "description": "delete with group_question",
                "operationId": "chapterdelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "formData",
                        "description": "uid delete",
                        "required": true,
                        "schema": {
                            "required": [
                                "uid"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "group_question_id",
                        "in": "formData",
                        "description": "group_question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "group_question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
        "/package/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "4.Package"
                ],
                "summary": "get all package",
                "description": "return package with take and skip",
                "operationId": "package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
        "/package/add": {
            "post": {
                "tags": [
                    "4.Package"
                ],
                "summary": "add new package",
                "description": "add new package",
                "operationId": "packageadd",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "folder_id",
                        "in": "formData",
                        "description": "folder id",
                        "required": true,
                        "schema": {
                            "required": [
                                "folder_id"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "translate",
                        "in": "formData",
                        "description": "translate json",
                        "required": true,
                        "schema": {
                            "required": [
                                "category_code"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "package_cost",
                        "in": "formData",
                        "description": "package_cost value",
                        "required": true,
                        "schema": {
                            "required": [
                                "package_cost"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
    },
...
}